Create README.md
The Law Offices of Scott Sobol is a personal injury law firm serving clients throughout South Florida and the Chicago, Illinois area. We understand that personal injury—whether it’s related to an auto accident, a slip and fall, or some other negligence, or is the result of a natural disaster—can be overwhelming. But it doesn’t have to be.